session timeouts in authentication provider
