+++ src/main/scala/uk/gov/hmrc/play/frontend/auth/AuthenticationProvider.scala
@@ -16,11 +16,13 @@
 
 package uk.gov.hmrc.play.frontend.auth
 
+import org.joda.time.{DateTime, DateTimeZone, Duration}
 import play.api.Logger
 import play.api.mvc.Results._
 import play.api.mvc._
 import uk.gov.hmrc.play.http.{HeaderCarrier, SessionKeys}
 import uk.gov.hmrc.play.http.logging.MdcLoggingExecutionContext.fromLoggingDetails
+
 import scala.concurrent._
 import play.api.Play
 import play.api.Play.current
@@ -35,6 +37,8 @@ object AuthenticationProviderIds {
 }
 
 trait AuthenticationProvider {
+  val timeoutSeconds = 900
+
   type FailureResult = Result
 
   def id: String
@@ -43,6 +47,10 @@ trait AuthenticationProvider {
 
   def redirectToLogin(implicit request: Request[_]): Future[Result]
 
+  def userNeedsNewSession(session: Session, now: () => DateTime): Boolean = {
+    extractTimestamp(session).fold(false)(hasExpired(now))
+  }
+
   def handleSessionTimeout(implicit request: Request[_]): Future[Result] = redirectToLogin
 
   def handleNotAuthenticated(implicit request: Request[_]): PartialFunction[UserCredentials, Future[Either[AuthContext, FailureResult]]]
@@ -50,6 +58,19 @@ trait AuthenticationProvider {
   def handleAuthenticated(implicit request: Request[_]): PartialFunction[UserCredentials, Future[Either[AuthContext, Result]]] = PartialFunction.empty
 
   implicit def hc(implicit request: Request[_]) = HeaderCarrier.fromHeadersAndSession(request.headers,Some(request.session) )
+
+  private def extractTimestamp(session: Session): Option[DateTime] = {
+    try {
+      session.get(SessionKeys.lastRequestTimestamp) map (t => new DateTime(t.toLong, DateTimeZone.UTC))
+    } catch {
+      case e: NumberFormatException => None
+    }
+  }
+
+  private def hasExpired(now: () => DateTime)(timestamp: DateTime): Boolean = {
+    val timeOfExpiry = timestamp plus Duration.standardSeconds(timeoutSeconds)
+    now() isAfter timeOfExpiry
+  }
 }
 
 trait GovernmentGateway extends AuthenticationProvider {
